CP-20677: implement gpumon server interface and manage idl server
This is wrapped into a functor to be able to acquire the interface on
launch and share it with rrdd.  Closure and cleanup are delegated to
the gpumon daemon loop.

Gpumon is modified to avoid killing the daemon when the library is loaded
correctly but RRDD fails for some reason.
The new signal handler is replacing the one in Process, however this was
only adding a log line that has been re-introduced by hand.

Note that the daemon (and the service) are still called `xcp-rrdd-gpumon`,
if we decide to change it I believe it should be the scope of a
different ticket.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>