CP-24185 add C binding get_pgpu_vgpu_compatibility2()
This adds an alternative binding to check compatibility of vGPU and pGPU
based on an interface that takes vGPU metadata:

val get_pgpu_vgpu_compatibility2:
  interface -> vgpu_metadata -> pgpu_metadata -> vgpu_compatibility_t

Compare against the existing binding:

val get_pgpu_vgpu_compatibility:
  interface -> vgpu_instance -> pgpu_metadata -> vgpu_compatibility_t

The existing binding takes a vgpu instance as parameter. This is not
available in the context of a VM resume. Hence the need to expose an
interface that take vgpu metadata.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>