CP-24185 refactor get_pgpu_vgpu_compatibility
This changes the implementation of get_pgpu_vgpu_compatibility by using
the new C binding get_pgpu_vgpu_compatibility2 and obtaining the vGPU
metadata for that call using the new C binding for get_vgpu_metadata.

The purpose of this is to remove the C binding for
get_pgpu_vgpu_compatibility.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>