CA-275983 make providing pGPU metadata version dependent
This commit does two things:

* add code to obtain and log the version of the driver providing pGPU
  meta data

* decide based on driver version whether to return pGPU meta data or
  not. No pGPU meta data is returned for drivers that don't support vGPU
  migration.

This is to support an assumption in the toolstack that the presence of
pGPU meta data means the pGPU supports migration. This is not the case
for early versions of NVDIA drivers.

For now all major version >= 385 are assumed to support migration but this
can be easily changed.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>