XSI-131 use correct versioned interface fn's
The NVIDIA interface redefines the names of interface functions:

  #define nvmlDeviceGetPciInfo        nvmlDeviceGetPciInfo_v3

When binding these functions dynamically in dlsym() we need to make sure
we use as a string the correct versioned name of the function. We
introduce the STR() macro to take care of this.

This fixes a number of problems: without the STR macro we were binding
the wrong version of interface functions. This caused in particular
nvmlDeviceGetPciInfo to fail silently. It provided no valid string in
the busId of nvmlPciInfo_t, which in turn lead to invalid UTF8 in the
RRD defined by gpumon. Subsequently this lead to failures of xapi when
the RRD data was streamed. See XSI-131 for details.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>