CP-33044 Introduce NVML.{attach,detach}
The NVML library is required to query Nvidia GPUs at runtime. This
library is loaded dynamically using dlopen(2). Add the capability to
attach and detach this library using the xapi-idl IPC mechanism. This is
experimental and replaced the exisitng mock implementation.

To avoid spamming logs when NVML is not available (on systems not using
NVidia), slow down retries and logging to once every 20min.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>